Canonical solvers in prog args (#59)
Add a factory class that names, describes, and creates canonical solvers. Add the option of selecting which solvers to run through a program argument, passing a list of canonical solver names.

Implemented with mostly basic solvers with default initializations. However, a big advantage is that now we will be able to define canonical ways to initialize various specific solvers. That way we will have a single source of truth for how to initialize a type of solver that we call X and is initialized by creating class Y with arguments Z. This will also force us to explicitly say how this changes when we later change, for example, the constructor of class Y.

(cherry picked from commit 7df947076a90c4d71a63ef560e327f02c73ae714)